cookiecutter/{{cookiecutter.project_name}}/requirements/develop.txt
wheel
watchdog
tox
flake8
twine
{% if cookiecutter.use_pytest == "yes" -%}
pytest-cov
{% else %}
nose
coverage
{% endif %}
{% if cookiecutter.use_paver == "yes" -%}
paver
{% endif %}
